Move `DesktopAudioInfo` to a more suitable place
